[backend] NFC: Split architecture dependant and independant parts of FMA dot conversion (#5655)
This PR splits FMA dot conversion from Triton GPU to LLVM in two parts:
- Common code with iteration across M/N dim
- Architecture dependant scalar multiplication of vectos across K dim

This PR do not introduce any test, because it does not fix any bugs or
introduce new functionality, it just refactors code.
